requirements.txt
django
python-dotenv
Pillow
requests
django-crispy-forms
gunicorn
django-heroku
boto3